🚢

+++ include/ble.hpp
@@ -8,5 +8,5 @@ void disconn();
 void parseBLE(char *);
 void pairBT(uint32_t);
 void setBLEBatteryLevel(uint8_t);
-void sendBLE(String, int = 1);
+void sendBLE(String, uint8_t = 1);
 void sendBLEf(const char *, ...);
